+++ esxi/esxi.main.go
@@ -1,0 +1,9 @@
+package esxi
+
+
+type SshConnectionInfo struct {
+    host string
+    port string
+    user string
+		pass string
+}
