--- esxi/ssh_connect.go
@@ -1,29 +1,0 @@
-package esxi
-
-import (
-	"golang.org/x/crypto/ssh"
-)
-
-
-// Connect to esxi host using ssh
-func connectToHost(user string, pass string, host string) (*ssh.Client, *ssh.Session, error) {
-
-	sshConfig := &ssh.ClientConfig{
-		User: user,
-		Auth: []ssh.AuthMethod{ssh.Password(pass)},
-	}
-	sshConfig.HostKeyCallback = ssh.InsecureIgnoreHostKey()
-
-	client, err := ssh.Dial("tcp", host, sshConfig)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	session, err := client.NewSession()
-	if err != nil {
-		client.Close()
-		return nil, nil, err
-	}
-
-	return client, session, nil
-}
